requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2==2.9.9
django-cors-headers==4.3.1
uwsgi==2.0.19.1
